api/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0
librosa>=0.10.2
torch>=2.3.0
numpy>=1.26.4
scipy>=1.13.0
pydantic>=2.7.0
python-multipart>=0.0.9
pytest>=8.2.0
pytest-asyncio>=0.23.7
aiofiles>=23.2.1
colorama>=0.4.6
google-genai>=1.0.0
transformers>=4.30.0
torchaudio>=2.0.0
soundfile>=0.12.1
